Remove redundant card handler module tests
